fix wrong cast in texture position

+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: averin <averin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 10:52:38 by averin            #+#    #+#             */
-/*   Updated: 2024/03/28 10:22:30 by averin           ###   ########.fr       */
+/*   Updated: 2024/03/28 10:42:51 by averin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -146,7 +146,7 @@ void	print_image(t_data *data)
 			return ;
 		}
 
-		int texX = (int) wallX * (double) tex->width;
+		int texX = wallX * (double) tex->width;
 		if ((face == 0 || face == 1) && direction.x > 0) texX = tex->width - texX - 1;
 		else if (direction.y < 0) texX = tex->width - texX - 1;
